Fixed GitHub username in README

+++ .local/bin/theme-menu.sh
@@ -4,12 +4,12 @@ WAYBAR_DIR="${HOME}/.config/waybar"
 STATE_FILE="$WAYBAR_DIR/.current"
 
 while true; do
-    CHOICE=$(printf "Waybar\nThemes\nCursors\nRefresh Rate\nResolution\nDefault Browser\nDefault File Manager\nSystem Reset" | rofi -dmenu -p "Settings" -theme-str 'configuration { show-icons: false; }')
+    CHOICE=$(printf " \uf233 Waybar\n \uf1fc Themes\n \uf245 Cursors\n \uf0e4  Refresh Rate\n \uf26c  Resolution\n \uf0ac  Default Browser\n \uf07b  Default File Manager\n \uf013  System Reset" | rofi -dmenu -p "Settings" -theme-str 'configuration { show-icons: false; }')
 
     [ -z "$CHOICE" ] && exit 0
 
     case "$CHOICE" in
-        Cursors)
+        *Cursors)
             CURSOR_STATE="$HOME/.config/hypr/.cursor-theme"
             current_cursor=$(cat "$CURSOR_STATE" 2>/dev/null)
 
@@ -48,12 +48,12 @@ while true; do
             CUR=$(echo "$CUR" | sed 's/ \[active\]$//')
             "$HOME/.local/bin/cursor-switcher.sh" "$CUR"
             ;;
-        Themes)
+        *Themes)
             THEME=$(printf "Dark\nLight" | rofi -dmenu -p "Theme" -theme-str 'configuration { show-icons: false; }')
             [ -z "$THEME" ] && continue
             "$HOME/.local/bin/theme-switcher" "$THEME"
             ;;
-        Waybar)
+        *Waybar)
             CONFIGS=()
             for f in "$WAYBAR_DIR"/config-*.jsonc; do
                 [ -f "$f" ] || continue
@@ -84,39 +84,35 @@ while true; do
             echo "$SEL" > "$STATE_FILE"
             "$WAYBAR_DIR/launch.sh"
             ;;
-        Refresh\ Rate)
+        *Refresh\ Rate)
             "$HOME/.local/bin/refresh-rate-menu.sh"
             ;;
-        Resolution)
+        *Resolution)
             "$HOME/.local/bin/resolution-menu.sh"
             ;;
-        Default\ Browser)
-            declare -A browser_map=(
-                ["brave-browser.desktop"]="Brave"
-                ["firefox.desktop"]="Firefox"
-                ["zen.desktop"]="Zen Browser"
-                ["org.torproject.torbrowser-launcher.desktop"]="Tor Browser"
-            )
-
+        *Default\ Browser)
             current_browser=$(xdg-settings get default-web-browser 2>/dev/null)
 
-            OPTIONS=("Brave" "Firefox" "Zen Browser" "Tor Browser")
-
             TMPFILE=$(mktemp)
-            for opt in "${OPTIONS[@]}"; do
-                desktop_file=""
-                for key in "${!browser_map[@]}"; do
-                    if [[ "${browser_map[$key]}" == "$opt" ]]; then
-                        desktop_file="$key"
-                        break
+            for dir in /usr/share/applications /usr/local/share/applications ~/.local/share/applications; do
+                [ -d "$dir" ] || continue
+                for f in "$dir"/*.desktop; do
+                    [ -f "$f" ] || continue
+                    grep -qi "x-scheme-handler/http" "$f" || continue
+                    grep -qi "NoDisplay=true" "$f" && continue
+                    name=$(grep "^Name=" "$f" 2>/dev/null | head -1 | cut -d= -f2-)
+                    [ -z "$name" ] && continue
+                    [[ "$name" == *"Settings"* ]] && continue
+                    desktop=$(basename "$f")
+                    if [[ "$desktop" == "$current_browser" ]]; then
+                        echo "${name} [active]" >> "$TMPFILE"
+                    else
+                        echo "$name" >> "$TMPFILE"
                     fi
                 done
-                if [[ "$desktop_file" == "$current_browser" ]]; then
-                    echo "${opt} [active]" >> "$TMPFILE"
-                else
-                    echo "$opt" >> "$TMPFILE"
-                fi
             done
+
+            [ ! -s "$TMPFILE" ] && { rm -f "$TMPFILE"; continue; }
 
             SEL=$(cat "$TMPFILE" | rofi -dmenu -p "Default Browser" -theme-str 'configuration { show-icons: false; }')
             rm -f "$TMPFILE"
@@ -125,40 +121,42 @@ while true; do
             SEL=$(echo "$SEL" | sed 's/ \[active\]$//')
 
             desktop_file=""
-            for key in "${!browser_map[@]}"; do
-                if [[ "${browser_map[$key]}" == "$SEL" ]]; then
-                    desktop_file="$key"
-                    break
-                fi
-            done
-
-            xdg-settings set default-web-browser "$desktop_file"
-            ;;
-        Default\ File\ Manager)
-            declare -A fm_map=(
-                ["thunar.desktop"]="Thunar"
-                ["yazi.desktop"]="Yazi"
-            )
-
-            current_fm=$(xdg-mime query default inode/directory 2>/dev/null)
-
-            OPTIONS=("Thunar" "Yazi")
-
-            TMPFILE=$(mktemp)
-            for opt in "${OPTIONS[@]}"; do
-                desktop_file=""
-                for key in "${!fm_map[@]}"; do
-                    if [[ "${fm_map[$key]}" == "$opt" ]]; then
-                        desktop_file="$key"
-                        break
+            for dir in /usr/share/applications /usr/local/share/applications ~/.local/share/applications; do
+                [ -d "$dir" ] || continue
+                for f in "$dir"/*.desktop; do
+                    [ -f "$f" ] || continue
+                    name=$(grep "^Name=" "$f" 2>/dev/null | head -1 | cut -d= -f2-)
+                    if [[ "$name" == "$SEL" ]]; then
+                        desktop_file=$(basename "$f")
+                        break 2
                     fi
                 done
-                if [[ "$desktop_file" == "$current_fm" ]]; then
-                    echo "${opt} [active]" >> "$TMPFILE"
-                else
-                    echo "$opt" >> "$TMPFILE"
-                fi
             done
+
+            [ -n "$desktop_file" ] && xdg-settings set default-web-browser "$desktop_file"
+            ;;
+        *Default\ File\ Manager)
+            current_fm=$(xdg-mime query default inode/directory 2>/dev/null)
+
+            TMPFILE=$(mktemp)
+            for dir in /usr/share/applications /usr/local/share/applications ~/.local/share/applications; do
+                [ -d "$dir" ] || continue
+                for f in "$dir"/*.desktop; do
+                    [ -f "$f" ] || continue
+                    grep -qi "^Categories=.*FileManager" "$f" || continue
+                    grep -qi "NoDisplay=true" "$f" && continue
+                    name=$(grep "^Name=" "$f" 2>/dev/null | head -1 | cut -d= -f2-)
+                    [ -z "$name" ] && continue
+                    desktop=$(basename "$f")
+                    if [[ "$desktop" == "$current_fm" ]]; then
+                        echo "${name} [active]" >> "$TMPFILE"
+                    else
+                        echo "$name" >> "$TMPFILE"
+                    fi
+                done
+            done
+
+            [ ! -s "$TMPFILE" ] && { rm -f "$TMPFILE"; continue; }
 
             SEL=$(cat "$TMPFILE" | rofi -dmenu -p "Default File Manager" -theme-str 'configuration { show-icons: false; }')
             rm -f "$TMPFILE"
@@ -167,16 +165,21 @@ while true; do
             SEL=$(echo "$SEL" | sed 's/ \[active\]$//')
 
             desktop_file=""
-            for key in "${!fm_map[@]}"; do
-                if [[ "${fm_map[$key]}" == "$SEL" ]]; then
-                    desktop_file="$key"
-                    break
-                fi
+            for dir in /usr/share/applications /usr/local/share/applications ~/.local/share/applications; do
+                [ -d "$dir" ] || continue
+                for f in "$dir"/*.desktop; do
+                    [ -f "$f" ] || continue
+                    name=$(grep "^Name=" "$f" 2>/dev/null | head -1 | cut -d= -f2-)
+                    if [[ "$name" == "$SEL" ]]; then
+                        desktop_file=$(basename "$f")
+                        break 2
+                    fi
+                done
             done
 
-            xdg-mime default "$desktop_file" inode/directory x-scheme-handler/file
+            [ -n "$desktop_file" ] && xdg-mime default "$desktop_file" inode/directory x-scheme-handler/file
             ;;
-        System\ Reset)
+        *System\ Reset)
             "$HOME/.local/bin/system-reset.sh"
             ;;
     esac
